Add custom link formatting callback
This replaces linkColor and linkOpacity with a single link(callback) method.

The callback is passed a d3 selection with the links, which can be styled in
exactly the same way as any other d3 object. This is more flexible and avoids
needing to learn the specific method names, although it does make the simple
case of changing the colour slightly harder.